requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt
psycopg[binary]
python-dotenv